Finish restructuring remaining components
