+++ static/src/components/App/index.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import LoginContainer from '../../containers/Login/LoginContainer';
 import MapContainer from '../../containers/Map/MapContainer';
 import SidebarContainer from '../../containers/Sidebar/SidebarContainer';
-// import MessagesBox from '../MessagesBox';
-// import Logo from '../Logo';
+import MessagesBoxContainer from '../../containers/MessagesBox/MessagesBoxContainer';
+import Logo from '../Logo';
 // import Modal from '../Modal';
 import './App.scss';
 
@@ -13,9 +13,9 @@ function App() {
       <LoginContainer />
       <MapContainer />
       <SidebarContainer />
-      {/** <MessagesBox />
+      <MessagesBoxContainer />
       <Logo />
-      <Modal /> */}
+      {/** <Modal /> */}
     </div>
   );
 }
