+++ static/src/components/Logo/index.jsx
@@ -1,27 +1,20 @@
 import React from 'react';
-import '../../stylesheets/Logo.scss';
+import './Logo.scss';
 
-class Logo extends React.Component {
-  shouldComponentUpdate() {
-    return false;
-  }
-  render() {
-    return (
-      <svg width="100%" height="100%" viewBox="0 0 400 70" className="fse-logo">
-        <g id="FS"><g><g>
-          <path
-            d={'M104.296,29.965c0.07,0.006 0.225,0.015 0.465,' +
-            '0.029l0,5.539c-0.342,-0.039 -0.645,-0.064 -0.91,-0.077c-0.266,' +
-            '-0.012 -0.481,-0.018 -0.645,-0.018c-2.175,0 -3.636,0.707 ' +
-            '-4.382,2.123c-0.417,0.797 -0.626,2.024 -0.626,3.68l0,' +
-            '9.883l-5.444,0l0,-20.676l5.159,0l0,3.604c0.835,-1.378 ' +
-            '1.562,-2.32 2.182,-2.826c1.011,-0.847 2.326,-1.271 3.945,' +
-            '-1.271c0.102,0 0.187,0.003 0.256,0.01'}
-            style={{ fill: '#fff', fillRule: 'nonzero' }}
-          />
-        <path d="M112.044,35.538c-0.765,0.786 -1.244,1.853 -1.441,3.2l9.314,0c-0.098,-1.435 -0.578,-2.524 -1.44,-3.268c-0.862,-0.742 -1.931,-1.113 -3.206,-1.113c-1.388,0 -2.464,0.394 -3.227,1.181M120.131,30.867c1.442,0.646 2.633,1.666 3.573,3.059c0.847,1.229 1.396,2.654 1.647,4.274c0.146,0.95 0.205,2.318 0.178,4.104l-15.058,0c0.084,2.074 0.804,3.528 2.162,4.362c0.826,0.519 1.819,0.778 2.981,0.778c1.231,0 2.232,-0.316 3.001,-0.948c0.42,-0.341 0.791,-0.816 1.113,-1.422l5.519,0c-0.145,1.226 -0.814,2.471 -2.004,3.736c-1.852,2.011 -4.445,3.016 -7.778,3.016c-2.752,0 -5.18,-0.848 -7.283,-2.544c-2.104,-1.695 -3.155,-4.455 -3.155,-8.278c0,-3.581 0.949,-6.328 2.848,-8.239c1.899,-1.911 4.364,-2.866 7.394,-2.866c1.799,0 3.42,0.322 4.862,0.968" style={{ fill: '#fff', fillRule: 'nonzero' }} /><path d="M153.758,44.523c0.115,0.961 0.363,1.644 0.744,2.048c0.673,0.721 1.918,1.082 3.735,1.082c1.067,0 1.916,-0.158 2.545,-0.474c0.628,-0.317 0.943,-0.79 0.943,-1.423c0,-0.607 -0.253,-1.068 -0.759,-1.385c-0.507,-0.316 -2.393,-0.859 -5.656,-1.631c-2.35,-0.582 -4.005,-1.309 -4.966,-2.182c-0.961,-0.859 -1.441,-2.098 -1.441,-3.718c0,-1.908 0.75,-3.549 2.252,-4.922c1.502,-1.371 3.615,-2.057 6.339,-2.057c2.586,0 4.693,0.515 6.321,1.546c1.628,1.031 2.563,2.81 2.804,5.339l-5.406,0c-0.076,-0.696 -0.272,-1.246 -0.59,-1.65c-0.596,-0.733 -1.611,-1.101 -3.043,-1.101c-1.179,0 -2.019,0.184 -2.52,0.551c-0.501,0.367 -0.752,0.796 -0.752,1.29c0,0.62 0.267,1.068 0.8,1.346c0.532,0.292 2.413,0.791 5.644,1.499c2.153,0.505 3.768,1.271 4.842,2.295c1.063,1.037 1.594,2.333 1.594,3.888c0,2.049 -0.764,3.721 -2.29,5.017c-1.528,1.297 -3.888,1.945 -7.08,1.945c-3.257,0 -5.661,-0.687 -7.213,-2.058c-1.552,-1.372 -2.329,-3.121 -2.329,-5.245l5.482,0Z" style={{ fill: '#a81921', fillRule: 'nonzero' }} /><path d="M182.069,45.607c0.847,-1.124 1.271,-2.721 1.271,-4.792c0,-2.071 -0.424,-3.665 -1.271,-4.782c-0.847,-1.117 -2.061,-1.677 -3.642,-1.677c-1.581,0 -2.797,0.56 -3.652,1.677c-0.853,1.117 -1.279,2.711 -1.279,4.782c0,2.071 0.426,3.668 1.279,4.792c0.855,1.123 2.071,1.685 3.652,1.685c1.581,0 2.795,-0.562 3.642,-1.685M186.394,48.594c-1.745,2.155 -4.395,3.232 -7.948,3.232c-3.554,0 -6.203,-1.077 -7.948,-3.232c-1.745,-2.154 -2.617,-4.747 -2.617,-7.779c0,-2.982 0.872,-5.565 2.617,-7.752c1.745,-2.185 4.394,-3.279 7.948,-3.279c3.553,0 6.203,1.094 7.948,3.279c1.745,2.187 2.618,4.77 2.618,7.752c0,3.032 -0.873,5.625 -2.618,7.779" style={{ fill: '#a81921', fillRule: 'nonzero' }} /><path d="M195.65,30.448l0,12.462c0,1.176 0.139,2.063 0.415,2.657c0.491,1.049 1.453,1.574 2.888,1.574c1.836,0 3.095,-0.746 3.775,-2.239c0.352,-0.809 0.528,-1.877 0.528,-3.205l0,-11.249l5.482,0l0,20.676l-5.254,0l0,-2.92c-0.051,0.062 -0.177,0.252 -0.379,0.568c-0.201,0.317 -0.441,0.595 -0.718,0.834c-0.845,0.759 -1.661,1.279 -2.449,1.556c-0.789,0.279 -1.712,0.418 -2.772,0.418c-3.051,0 -5.107,-1.101 -6.166,-3.301c-0.593,-1.215 -0.889,-3.003 -0.889,-5.369l0,-12.462l5.539,0Z" style={{ fill: '#a81921', fillRule: 'nonzero' }} /><path d="M227.298,31.634c1.354,1.119 2.031,2.975 2.031,5.568l0,13.922l-5.539,0l0,-12.576c0,-1.087 -0.144,-1.922 -0.433,-2.504c-0.527,-1.063 -1.532,-1.593 -3.013,-1.593c-1.82,0 -3.069,0.778 -3.747,2.332c-0.352,0.823 -0.527,1.873 -0.527,3.15l0,11.191l-5.406,0l0,-20.637l5.235,0l0,3.015c0.693,-1.062 1.348,-1.827 1.965,-2.295c1.107,-0.835 2.512,-1.252 4.212,-1.252c2.128,0 3.869,0.56 5.222,1.679" style={{ fill: '#a81921', fillRule: 'nonzero' }} /><path d="M244.179,45.509c0.796,-1.138 1.194,-2.611 1.194,-4.419c0,-2.529 -0.638,-4.337 -1.915,-5.426c-0.785,-0.657 -1.695,-0.986 -2.732,-0.986c-1.581,0 -2.741,0.597 -3.481,1.793c-0.739,1.194 -1.109,2.678 -1.109,4.449c0,1.909 0.376,3.435 1.128,4.579c0.753,1.145 1.894,1.717 3.425,1.717c1.529,0 2.693,-0.568 3.49,-1.707M242.85,30.781c0.974,0.55 1.765,1.312 2.372,2.285l0,-9.863l5.482,0l0,27.921l-5.255,0l0,-2.864c-0.771,1.226 -1.65,2.118 -2.636,2.674c-0.987,0.557 -2.213,0.835 -3.68,0.835c-2.416,0 -4.449,-0.977 -6.099,-2.93c-1.65,-1.954 -2.475,-4.461 -2.475,-7.522c0,-3.527 0.813,-6.303 2.437,-8.327c1.625,-2.023 3.797,-3.035 6.516,-3.035c1.252,0 2.365,0.275 3.338,0.826" style={{ fill: '#a81921', fillRule: 'nonzero' }} /><path d="M133.874,35.538c-0.764,0.786 -1.244,1.853 -1.441,3.2l9.314,0c-0.098,-1.435 -0.578,-2.524 -1.44,-3.268c-0.862,-0.742 -1.931,-1.113 -3.206,-1.113c-1.388,0 -2.463,0.394 -3.227,1.181M141.961,30.867c1.442,0.646 2.634,1.666 3.573,3.059c0.847,1.229 1.396,2.654 1.647,4.274c0.146,0.95 0.205,2.318 0.178,4.104l-15.058,0c0.084,2.074 0.805,3.528 2.162,4.362c0.826,0.519 1.819,0.778 2.981,0.778c1.231,0 2.232,-0.316 3.001,-0.948c0.42,-0.341 0.791,-0.816 1.113,-1.422l5.52,0c-0.146,1.226 -0.814,2.471 -2.005,3.736c-1.852,2.011 -4.445,3.016 -7.778,3.016c-2.752,0 -5.18,-0.848 -7.283,-2.544c-2.104,-1.695 -3.156,-4.455 -3.156,-8.278c0,-3.581 0.95,-6.328 2.849,-8.239c1.899,-1.911 4.363,-2.866 7.394,-2.866c1.799,0 3.42,0.322 4.862,0.968" style={{ fill: '#fff', fillRule: 'nonzero' }} /><path d="M88.173,30.638l0,-1.745c0,-0.404 0.136,-0.775 0.408,-1.11c0.272,-0.334 0.765,-0.493 1.479,-0.474c0.714,0.019 1.224,0.048 1.527,0.085l0,-4.401c-0.48,-0.037 -0.86,-0.066 -1.137,-0.085c-0.279,-0.019 -0.589,-0.028 -0.93,-0.028c-2.997,0 -4.888,0.619 -5.671,1.859c-0.747,0.834 -1.12,2.359 -1.12,4.571l0,1.319l-5.703,0c-1.29,0.009 -2.473,0.437 -3.515,1.071c-1.036,0.646 -1.948,1.511 -2.535,2.66l-0.458,0.912l-1.167,-4.766c-0.132,-0.522 -0.28,-0.974 -0.564,-1.448c-0.148,-0.237 -0.335,-0.49 -0.651,-0.738c-0.307,-0.246 -0.806,-0.478 -1.367,-0.475c-0.559,-0.003 -1.054,0.227 -1.36,0.467c-0.878,0.759 -0.984,1.392 -1.228,2.167l-1.817,7.091l-2.693,-19.688c-0.069,-0.487 -0.142,-0.879 -0.285,-1.296c-0.078,-0.213 -0.161,-0.434 -0.384,-0.74c-0.115,-0.151 -0.272,-0.333 -0.538,-0.512c-0.261,-0.176 -0.669,-0.337 -1.112,-0.333c-0.44,-0.002 -0.842,0.155 -1.1,0.327c-0.457,0.311 -0.609,0.593 -0.737,0.805c-0.321,0.626 -0.382,1.062 -0.497,1.71l-4.508,29.155l-1.736,-23.979c-0.037,-0.477 -0.079,-0.854 -0.171,-1.242c-0.052,-0.201 -0.1,-0.4 -0.266,-0.711c-0.09,-0.155 -0.209,-0.352 -0.472,-0.582c-0.249,-0.225 -0.734,-0.487 -1.294,-0.482c-0.552,-0.002 -0.993,0.235 -1.234,0.431c-0.428,0.36 -0.521,0.609 -0.628,0.804c-0.25,0.563 -0.312,0.968 -0.421,1.592l-1.858,11.716l-2.879,-27.371c-0.051,-0.471 -0.105,-0.843 -0.206,-1.223c-0.057,-0.195 -0.109,-0.387 -0.275,-0.685c-0.089,-0.149 -0.206,-0.337 -0.461,-0.559c-0.243,-0.215 -0.715,-0.481 -1.288,-0.477c-0.57,-0.003 -1.034,0.256 -1.276,0.465c-0.432,0.388 -0.51,0.643 -0.611,0.843c-0.226,0.571 -0.272,0.975 -0.358,1.606l-4.938,41.268l-2.721,-21.228c-0.068,-0.499 -0.14,-0.907 -0.301,-1.35c-0.088,-0.224 -0.184,-0.463 -0.433,-0.775c-0.224,-0.289 -0.777,-0.777 -1.614,-0.782c-0.813,0.01 -1.323,0.446 -1.554,0.707c-0.366,0.426 -0.464,0.725 -0.586,1.038c-0.109,0.311 -0.189,0.629 -0.264,0.992l-2.353,11.603l-1.686,-4.046c-0.511,-1.203 -1.385,-2.125 -2.398,-2.808c-1.021,-0.673 -2.208,-1.125 -3.516,-1.134l-6.188,0c-1.058,0 -1.915,0.857 -1.915,1.915c0,1.057 0.857,1.915 1.915,1.915l6.188,0c0.34,-0.009 0.91,0.162 1.391,0.491c0.49,0.318 0.866,0.777 0.989,1.094l4.089,9.81c0.322,0.774 1.119,1.251 1.953,1.169c0.834,-0.081 1.524,-0.703 1.69,-1.525l1.887,-9.309l3.213,25.067c0.062,0.474 0.126,0.852 0.246,1.242c0.066,0.201 0.131,0.403 0.319,0.703c0.099,0.148 0.232,0.333 0.493,0.537c0.251,0.197 0.696,0.416 1.213,0.413c0.518,0.003 0.967,-0.216 1.219,-0.417c0.448,-0.368 0.55,-0.64 0.658,-0.846c0.255,-0.599 0.299,-1.013 0.388,-1.651l4.751,-39.703l2.617,24.881c0.053,0.48 0.11,0.867 0.23,1.273c0.065,0.207 0.131,0.421 0.329,0.732c0.104,0.154 0.246,0.347 0.515,0.549c0.259,0.198 0.701,0.398 1.191,0.394c0.484,0.002 0.904,-0.186 1.155,-0.369c0.447,-0.332 0.574,-0.601 0.691,-0.805c0.29,-0.597 0.352,-1.018 0.467,-1.654l1.532,-9.662l1.757,24.262c0.036,0.475 0.077,0.85 0.168,1.237c0.051,0.198 0.098,0.395 0.261,0.704c0.088,0.155 0.204,0.351 0.465,0.584c0.244,0.224 0.724,0.491 1.288,0.491l0.017,0c0.562,0.001 1.008,-0.244 1.247,-0.443c0.426,-0.365 0.513,-0.611 0.618,-0.807c0.243,-0.558 0.304,-0.96 0.41,-1.584l5.083,-32.865l2.238,16.358c0.073,0.513 0.158,0.945 0.353,1.417c0.105,0.237 0.23,0.498 0.515,0.806c0.264,0.293 0.831,0.677 1.542,0.673c0.682,-0.002 1.183,-0.327 1.45,-0.576c0.412,-0.393 0.555,-0.712 0.704,-1.03c0.139,-0.318 0.242,-0.644 0.336,-1.009l2.347,-9.162l0.586,2.396c0.15,0.593 0.352,1.138 0.753,1.667c0.374,0.514 1.14,1.071 2.066,1.062c0.815,-0.004 1.446,-0.386 1.854,-0.769c0.42,-0.395 0.705,-0.834 0.954,-1.323l1.448,-2.883c0.152,-0.322 0.585,-0.794 1.122,-1.119c0.531,-0.336 1.15,-0.51 1.506,-0.501l5.779,0l0,16.665l5.368,0l0,-16.673l3.547,0l0,-3.813l-3.547,0Z" style={{ fill:'#fff', fillRule: 'nonzero' }}/></g><path d="M263.529,29.929c-3.146,0 -5.638,1.07 -7.436,3.251c-1.88,2.14 -2.779,4.773 -2.779,7.942c0,3.539 0.981,6.296 2.942,8.312c1.839,1.934 4.331,2.922 7.478,2.922c2.819,0 5.148,-0.823 7.028,-2.428c1.471,-1.317 2.452,-3.004 2.942,-4.979l-4.658,0c-0.572,1.193 -1.185,2.057 -1.88,2.592c-0.899,0.658 -2.043,0.988 -3.473,0.988c-1.675,0 -2.983,-0.535 -3.882,-1.564c-0.899,-1.029 -1.43,-2.551 -1.594,-4.527l15.814,0c0,-3.826 -0.858,-6.83 -2.533,-8.97c-1.839,-2.387 -4.495,-3.539 -7.969,-3.539ZM263.652,33.756c3.269,0 5.108,1.687 5.516,5.144l-10.869,0c0.286,-1.688 0.858,-2.963 1.716,-3.827c0.899,-0.906 2.084,-1.317 3.637,-1.317Z" style={{ fill:'#535353', fillRule: 'nonzero' }}/><path d="M274.521,30.505l7.274,9.958l-8.213,11.316l5.475,0l5.312,-7.859l5.271,7.859l5.476,0l-8.213,-11.316l7.273,-9.958l-5.435,0l-4.372,6.502l-4.413,-6.502l-5.435,0Z" style={{ fill:'#535353', fillRule: 'nonzero' }}/><path d="M306.598,29.929c-2.779,0 -4.903,0.946 -6.374,2.839l0,-2.263l-4.332,0l0,29.422l4.658,0l0,-11.069c1.717,2.304 3.841,3.498 6.293,3.498c3.024,0 5.435,-1.111 7.151,-3.334c1.594,-2.057 2.411,-4.649 2.411,-7.777c0,-3.292 -0.817,-5.925 -2.452,-7.983c-1.798,-2.222 -4.25,-3.333 -7.355,-3.333ZM305.903,33.715c2.043,0 3.555,0.74 4.536,2.222c0.776,1.234 1.185,3.004 1.185,5.308c0,2.305 -0.449,4.074 -1.308,5.308c-0.98,1.317 -2.533,2.017 -4.576,2.017c-1.512,0 -2.738,-0.617 -3.719,-1.852c-1.103,-1.358 -1.634,-3.168 -1.634,-5.391l0,-0.329c0,-2.14 0.409,-3.827 1.308,-5.102c0.98,-1.482 2.37,-2.181 4.208,-2.181Z" style={{ fill:'#535353', fillRule: 'nonzero' }}/><rect x="319.02" y="21.822" width="4.658" height="29.957" style={{ fill:'#535353', fillRule: 'nonzero' }}/><path d="M336.918,29.929c-3.229,0 -5.803,1.029 -7.723,3.168c-1.962,2.099 -2.902,4.774 -2.902,8.066c0,3.251 0.94,5.925 2.861,7.983c1.961,2.14 4.535,3.21 7.764,3.21c3.187,0 5.802,-1.07 7.763,-3.21c1.88,-2.058 2.861,-4.732 2.861,-7.983c0,-3.292 -0.981,-5.967 -2.902,-8.066c-1.961,-2.139 -4.535,-3.168 -7.722,-3.168ZM336.918,33.756c1.92,0 3.432,0.741 4.494,2.304c0.899,1.317 1.39,3.045 1.39,5.103c0,2.016 -0.491,3.703 -1.39,5.061c-1.062,1.523 -2.574,2.305 -4.494,2.305c-1.962,0 -3.433,-0.782 -4.495,-2.305c-0.899,-1.317 -1.349,-3.004 -1.349,-5.061c0,-2.058 0.45,-3.786 1.349,-5.103c1.062,-1.563 2.533,-2.304 4.495,-2.304Z" style={{ fill:'#535353', fillRule: 'nonzero' }}/><path d="M359.719,29.929c-1.308,0 -2.411,0.37 -3.351,1.152c-0.777,0.576 -1.43,1.399 -2.002,2.469l0,-3.045l-4.618,0l0,21.274l4.618,0l0,-11.275c0,-1.81 0.531,-3.25 1.675,-4.361c0.981,-0.988 2.084,-1.441 3.31,-1.441c0.98,0 1.961,0.124 3.024,0.453l0,-4.691c-0.736,-0.37 -1.635,-0.535 -2.656,-0.535Z" style={{ fill:'#535353', fillRule: 'nonzero' }}/><path d="M372.141,29.929c-3.147,0 -5.639,1.07 -7.437,3.251c-1.88,2.14 -2.779,4.773 -2.779,7.942c0,3.539 0.981,6.296 2.942,8.312c1.839,1.934 4.331,2.922 7.478,2.922c2.819,0 5.148,-0.823 7.028,-2.428c1.471,-1.317 2.452,-3.004 2.942,-4.979l-4.658,0c-0.572,1.193 -1.185,2.057 -1.88,2.592c-0.899,0.658 -2.043,0.988 -3.473,0.988c-1.675,0 -2.983,-0.535 -3.882,-1.564c-0.899,-1.029 -1.43,-2.551 -1.594,-4.527l15.814,0c0,-3.826 -0.858,-6.83 -2.533,-8.97c-1.839,-2.387 -4.495,-3.539 -7.968,-3.539ZM372.263,33.756c3.269,0 5.108,1.687 5.517,5.144l-10.87,0c0.286,-1.688 0.858,-2.963 1.716,-3.827c0.899,-0.906 2.084,-1.317 3.637,-1.317Z" style={{ fill:'#535353', fillRule: 'nonzero' }}/><path d="M394.41,29.929c-1.307,0 -2.411,0.37 -3.35,1.152c-0.777,0.576 -1.431,1.399 -2.003,2.469l0,-3.045l-4.617,0l0,21.274l4.617,0l0,-11.275c0,-1.81 0.532,-3.25 1.676,-4.361c0.98,-0.988 2.084,-1.441 3.31,-1.441c0.98,0 1.961,0.124 3.023,0.453l0,-4.691c-0.735,-0.37 -1.634,-0.535 -2.656,-0.535Z" style={{ fill:'#535353', fillRule: 'nonzero' }}/></g></g></svg>
-      );
-  }
-}
+const Logo = () => (
+  <svg width="100%" height="100%" viewBox="0 0 400 70" className="fse-logo">
+    <g id="FS"><g><g>
+      <path
+        d={'M104.296,29.965c0.07,0.006 0.225,0.015 0.465,' +
+        '0.029l0,5.539c-0.342,-0.039 -0.645,-0.064 -0.91,-0.077c-0.266,' +
+        '-0.012 -0.481,-0.018 -0.645,-0.018c-2.175,0 -3.636,0.707 ' +
+        '-4.382,2.123c-0.417,0.797 -0.626,2.024 -0.626,3.68l0,' +
+        '9.883l-5.444,0l0,-20.676l5.159,0l0,3.604c0.835,-1.378 ' +
+        '1.562,-2.32 2.182,-2.826c1.011,-0.847 2.326,-1.271 3.945,' +
+        '-1.271c0.102,0 0.187,0.003 0.256,0.01'}
+        style={{ fill: '#fff', fillRule: 'nonzero' }}
+      />
+      <path d="M112.044,35.538c-0.765,0.786 -1.244,1.853 -1.441,3.2l9.314,0c-0.098,-1.435 -0.578,-2.524 -1.44,-3.268c-0.862,-0.742 -1.931,-1.113 -3.206,-1.113c-1.388,0 -2.464,0.394 -3.227,1.181M120.131,30.867c1.442,0.646 2.633,1.666 3.573,3.059c0.847,1.229 1.396,2.654 1.647,4.274c0.146,0.95 0.205,2.318 0.178,4.104l-15.058,0c0.084,2.074 0.804,3.528 2.162,4.362c0.826,0.519 1.819,0.778 2.981,0.778c1.231,0 2.232,-0.316 3.001,-0.948c0.42,-0.341 0.791,-0.816 1.113,-1.422l5.519,0c-0.145,1.226 -0.814,2.471 -2.004,3.736c-1.852,2.011 -4.445,3.016 -7.778,3.016c-2.752,0 -5.18,-0.848 -7.283,-2.544c-2.104,-1.695 -3.155,-4.455 -3.155,-8.278c0,-3.581 0.949,-6.328 2.848,-8.239c1.899,-1.911 4.364,-2.866 7.394,-2.866c1.799,0 3.42,0.322 4.862,0.968" style={{ fill: '#fff', fillRule: 'nonzero' }} /><path d="M153.758,44.523c0.115,0.961 0.363,1.644 0.744,2.048c0.673,0.721 1.918,1.082 3.735,1.082c1.067,0 1.916,-0.158 2.545,-0.474c0.628,-0.317 0.943,-0.79 0.943,-1.423c0,-0.607 -0.253,-1.068 -0.759,-1.385c-0.507,-0.316 -2.393,-0.859 -5.656,-1.631c-2.35,-0.582 -4.005,-1.309 -4.966,-2.182c-0.961,-0.859 -1.441,-2.098 -1.441,-3.718c0,-1.908 0.75,-3.549 2.252,-4.922c1.502,-1.371 3.615,-2.057 6.339,-2.057c2.586,0 4.693,0.515 6.321,1.546c1.628,1.031 2.563,2.81 2.804,5.339l-5.406,0c-0.076,-0.696 -0.272,-1.246 -0.59,-1.65c-0.596,-0.733 -1.611,-1.101 -3.043,-1.101c-1.179,0 -2.019,0.184 -2.52,0.551c-0.501,0.367 -0.752,0.796 -0.752,1.29c0,0.62 0.267,1.068 0.8,1.346c0.532,0.292 2.413,0.791 5.644,1.499c2.153,0.505 3.768,1.271 4.842,2.295c1.063,1.037 1.594,2.333 1.594,3.888c0,2.049 -0.764,3.721 -2.29,5.017c-1.528,1.297 -3.888,1.945 -7.08,1.945c-3.257,0 -5.661,-0.687 -7.213,-2.058c-1.552,-1.372 -2.329,-3.121 -2.329,-5.245l5.482,0Z" style={{ fill: '#a81921', fillRule: 'nonzero' }} /><path d="M182.069,45.607c0.847,-1.124 1.271,-2.721 1.271,-4.792c0,-2.071 -0.424,-3.665 -1.271,-4.782c-0.847,-1.117 -2.061,-1.677 -3.642,-1.677c-1.581,0 -2.797,0.56 -3.652,1.677c-0.853,1.117 -1.279,2.711 -1.279,4.782c0,2.071 0.426,3.668 1.279,4.792c0.855,1.123 2.071,1.685 3.652,1.685c1.581,0 2.795,-0.562 3.642,-1.685M186.394,48.594c-1.745,2.155 -4.395,3.232 -7.948,3.232c-3.554,0 -6.203,-1.077 -7.948,-3.232c-1.745,-2.154 -2.617,-4.747 -2.617,-7.779c0,-2.982 0.872,-5.565 2.617,-7.752c1.745,-2.185 4.394,-3.279 7.948,-3.279c3.553,0 6.203,1.094 7.948,3.279c1.745,2.187 2.618,4.77 2.618,7.752c0,3.032 -0.873,5.625 -2.618,7.779" style={{ fill: '#a81921', fillRule: 'nonzero' }} /><path d="M195.65,30.448l0,12.462c0,1.176 0.139,2.063 0.415,2.657c0.491,1.049 1.453,1.574 2.888,1.574c1.836,0 3.095,-0.746 3.775,-2.239c0.352,-0.809 0.528,-1.877 0.528,-3.205l0,-11.249l5.482,0l0,20.676l-5.254,0l0,-2.92c-0.051,0.062 -0.177,0.252 -0.379,0.568c-0.201,0.317 -0.441,0.595 -0.718,0.834c-0.845,0.759 -1.661,1.279 -2.449,1.556c-0.789,0.279 -1.712,0.418 -2.772,0.418c-3.051,0 -5.107,-1.101 -6.166,-3.301c-0.593,-1.215 -0.889,-3.003 -0.889,-5.369l0,-12.462l5.539,0Z" style={{ fill: '#a81921', fillRule: 'nonzero' }} /><path d="M227.298,31.634c1.354,1.119 2.031,2.975 2.031,5.568l0,13.922l-5.539,0l0,-12.576c0,-1.087 -0.144,-1.922 -0.433,-2.504c-0.527,-1.063 -1.532,-1.593 -3.013,-1.593c-1.82,0 -3.069,0.778 -3.747,2.332c-0.352,0.823 -0.527,1.873 -0.527,3.15l0,11.191l-5.406,0l0,-20.637l5.235,0l0,3.015c0.693,-1.062 1.348,-1.827 1.965,-2.295c1.107,-0.835 2.512,-1.252 4.212,-1.252c2.128,0 3.869,0.56 5.222,1.679" style={{ fill: '#a81921', fillRule: 'nonzero' }} /><path d="M244.179,45.509c0.796,-1.138 1.194,-2.611 1.194,-4.419c0,-2.529 -0.638,-4.337 -1.915,-5.426c-0.785,-0.657 -1.695,-0.986 -2.732,-0.986c-1.581,0 -2.741,0.597 -3.481,1.793c-0.739,1.194 -1.109,2.678 -1.109,4.449c0,1.909 0.376,3.435 1.128,4.579c0.753,1.145 1.894,1.717 3.425,1.717c1.529,0 2.693,-0.568 3.49,-1.707M242.85,30.781c0.974,0.55 1.765,1.312 2.372,2.285l0,-9.863l5.482,0l0,27.921l-5.255,0l0,-2.864c-0.771,1.226 -1.65,2.118 -2.636,2.674c-0.987,0.557 -2.213,0.835 -3.68,0.835c-2.416,0 -4.449,-0.977 -6.099,-2.93c-1.65,-1.954 -2.475,-4.461 -2.475,-7.522c0,-3.527 0.813,-6.303 2.437,-8.327c1.625,-2.023 3.797,-3.035 6.516,-3.035c1.252,0 2.365,0.275 3.338,0.826" style={{ fill: '#a81921', fillRule: 'nonzero' }} /><path d="M133.874,35.538c-0.764,0.786 -1.244,1.853 -1.441,3.2l9.314,0c-0.098,-1.435 -0.578,-2.524 -1.44,-3.268c-0.862,-0.742 -1.931,-1.113 -3.206,-1.113c-1.388,0 -2.463,0.394 -3.227,1.181M141.961,30.867c1.442,0.646 2.634,1.666 3.573,3.059c0.847,1.229 1.396,2.654 1.647,4.274c0.146,0.95 0.205,2.318 0.178,4.104l-15.058,0c0.084,2.074 0.805,3.528 2.162,4.362c0.826,0.519 1.819,0.778 2.981,0.778c1.231,0 2.232,-0.316 3.001,-0.948c0.42,-0.341 0.791,-0.816 1.113,-1.422l5.52,0c-0.146,1.226 -0.814,2.471 -2.005,3.736c-1.852,2.011 -4.445,3.016 -7.778,3.016c-2.752,0 -5.18,-0.848 -7.283,-2.544c-2.104,-1.695 -3.156,-4.455 -3.156,-8.278c0,-3.581 0.95,-6.328 2.849,-8.239c1.899,-1.911 4.363,-2.866 7.394,-2.866c1.799,0 3.42,0.322 4.862,0.968" style={{ fill: '#fff', fillRule: 'nonzero' }} /><path d="M88.173,30.638l0,-1.745c0,-0.404 0.136,-0.775 0.408,-1.11c0.272,-0.334 0.765,-0.493 1.479,-0.474c0.714,0.019 1.224,0.048 1.527,0.085l0,-4.401c-0.48,-0.037 -0.86,-0.066 -1.137,-0.085c-0.279,-0.019 -0.589,-0.028 -0.93,-0.028c-2.997,0 -4.888,0.619 -5.671,1.859c-0.747,0.834 -1.12,2.359 -1.12,4.571l0,1.319l-5.703,0c-1.29,0.009 -2.473,0.437 -3.515,1.071c-1.036,0.646 -1.948,1.511 -2.535,2.66l-0.458,0.912l-1.167,-4.766c-0.132,-0.522 -0.28,-0.974 -0.564,-1.448c-0.148,-0.237 -0.335,-0.49 -0.651,-0.738c-0.307,-0.246 -0.806,-0.478 -1.367,-0.475c-0.559,-0.003 -1.054,0.227 -1.36,0.467c-0.878,0.759 -0.984,1.392 -1.228,2.167l-1.817,7.091l-2.693,-19.688c-0.069,-0.487 -0.142,-0.879 -0.285,-1.296c-0.078,-0.213 -0.161,-0.434 -0.384,-0.74c-0.115,-0.151 -0.272,-0.333 -0.538,-0.512c-0.261,-0.176 -0.669,-0.337 -1.112,-0.333c-0.44,-0.002 -0.842,0.155 -1.1,0.327c-0.457,0.311 -0.609,0.593 -0.737,0.805c-0.321,0.626 -0.382,1.062 -0.497,1.71l-4.508,29.155l-1.736,-23.979c-0.037,-0.477 -0.079,-0.854 -0.171,-1.242c-0.052,-0.201 -0.1,-0.4 -0.266,-0.711c-0.09,-0.155 -0.209,-0.352 -0.472,-0.582c-0.249,-0.225 -0.734,-0.487 -1.294,-0.482c-0.552,-0.002 -0.993,0.235 -1.234,0.431c-0.428,0.36 -0.521,0.609 -0.628,0.804c-0.25,0.563 -0.312,0.968 -0.421,1.592l-1.858,11.716l-2.879,-27.371c-0.051,-0.471 -0.105,-0.843 -0.206,-1.223c-0.057,-0.195 -0.109,-0.387 -0.275,-0.685c-0.089,-0.149 -0.206,-0.337 -0.461,-0.559c-0.243,-0.215 -0.715,-0.481 -1.288,-0.477c-0.57,-0.003 -1.034,0.256 -1.276,0.465c-0.432,0.388 -0.51,0.643 -0.611,0.843c-0.226,0.571 -0.272,0.975 -0.358,1.606l-4.938,41.268l-2.721,-21.228c-0.068,-0.499 -0.14,-0.907 -0.301,-1.35c-0.088,-0.224 -0.184,-0.463 -0.433,-0.775c-0.224,-0.289 -0.777,-0.777 -1.614,-0.782c-0.813,0.01 -1.323,0.446 -1.554,0.707c-0.366,0.426 -0.464,0.725 -0.586,1.038c-0.109,0.311 -0.189,0.629 -0.264,0.992l-2.353,11.603l-1.686,-4.046c-0.511,-1.203 -1.385,-2.125 -2.398,-2.808c-1.021,-0.673 -2.208,-1.125 -3.516,-1.134l-6.188,0c-1.058,0 -1.915,0.857 -1.915,1.915c0,1.057 0.857,1.915 1.915,1.915l6.188,0c0.34,-0.009 0.91,0.162 1.391,0.491c0.49,0.318 0.866,0.777 0.989,1.094l4.089,9.81c0.322,0.774 1.119,1.251 1.953,1.169c0.834,-0.081 1.524,-0.703 1.69,-1.525l1.887,-9.309l3.213,25.067c0.062,0.474 0.126,0.852 0.246,1.242c0.066,0.201 0.131,0.403 0.319,0.703c0.099,0.148 0.232,0.333 0.493,0.537c0.251,0.197 0.696,0.416 1.213,0.413c0.518,0.003 0.967,-0.216 1.219,-0.417c0.448,-0.368 0.55,-0.64 0.658,-0.846c0.255,-0.599 0.299,-1.013 0.388,-1.651l4.751,-39.703l2.617,24.881c0.053,0.48 0.11,0.867 0.23,1.273c0.065,0.207 0.131,0.421 0.329,0.732c0.104,0.154 0.246,0.347 0.515,0.549c0.259,0.198 0.701,0.398 1.191,0.394c0.484,0.002 0.904,-0.186 1.155,-0.369c0.447,-0.332 0.574,-0.601 0.691,-0.805c0.29,-0.597 0.352,-1.018 0.467,-1.654l1.532,-9.662l1.757,24.262c0.036,0.475 0.077,0.85 0.168,1.237c0.051,0.198 0.098,0.395 0.261,0.704c0.088,0.155 0.204,0.351 0.465,0.584c0.244,0.224 0.724,0.491 1.288,0.491l0.017,0c0.562,0.001 1.008,-0.244 1.247,-0.443c0.426,-0.365 0.513,-0.611 0.618,-0.807c0.243,-0.558 0.304,-0.96 0.41,-1.584l5.083,-32.865l2.238,16.358c0.073,0.513 0.158,0.945 0.353,1.417c0.105,0.237 0.23,0.498 0.515,0.806c0.264,0.293 0.831,0.677 1.542,0.673c0.682,-0.002 1.183,-0.327 1.45,-0.576c0.412,-0.393 0.555,-0.712 0.704,-1.03c0.139,-0.318 0.242,-0.644 0.336,-1.009l2.347,-9.162l0.586,2.396c0.15,0.593 0.352,1.138 0.753,1.667c0.374,0.514 1.14,1.071 2.066,1.062c0.815,-0.004 1.446,-0.386 1.854,-0.769c0.42,-0.395 0.705,-0.834 0.954,-1.323l1.448,-2.883c0.152,-0.322 0.585,-0.794 1.122,-1.119c0.531,-0.336 1.15,-0.51 1.506,-0.501l5.779,0l0,16.665l5.368,0l0,-16.673l3.547,0l0,-3.813l-3.547,0Z" style={{ fill: '#fff', fillRule: 'nonzero' }} /></g><path d="M263.529,29.929c-3.146,0 -5.638,1.07 -7.436,3.251c-1.88,2.14 -2.779,4.773 -2.779,7.942c0,3.539 0.981,6.296 2.942,8.312c1.839,1.934 4.331,2.922 7.478,2.922c2.819,0 5.148,-0.823 7.028,-2.428c1.471,-1.317 2.452,-3.004 2.942,-4.979l-4.658,0c-0.572,1.193 -1.185,2.057 -1.88,2.592c-0.899,0.658 -2.043,0.988 -3.473,0.988c-1.675,0 -2.983,-0.535 -3.882,-1.564c-0.899,-1.029 -1.43,-2.551 -1.594,-4.527l15.814,0c0,-3.826 -0.858,-6.83 -2.533,-8.97c-1.839,-2.387 -4.495,-3.539 -7.969,-3.539ZM263.652,33.756c3.269,0 5.108,1.687 5.516,5.144l-10.869,0c0.286,-1.688 0.858,-2.963 1.716,-3.827c0.899,-0.906 2.084,-1.317 3.637,-1.317Z" style={{ fill: '#535353', fillRule: 'nonzero' }} /><path d="M274.521,30.505l7.274,9.958l-8.213,11.316l5.475,0l5.312,-7.859l5.271,7.859l5.476,0l-8.213,-11.316l7.273,-9.958l-5.435,0l-4.372,6.502l-4.413,-6.502l-5.435,0Z" style={{ fill: '#535353', fillRule: 'nonzero' }} /><path d="M306.598,29.929c-2.779,0 -4.903,0.946 -6.374,2.839l0,-2.263l-4.332,0l0,29.422l4.658,0l0,-11.069c1.717,2.304 3.841,3.498 6.293,3.498c3.024,0 5.435,-1.111 7.151,-3.334c1.594,-2.057 2.411,-4.649 2.411,-7.777c0,-3.292 -0.817,-5.925 -2.452,-7.983c-1.798,-2.222 -4.25,-3.333 -7.355,-3.333ZM305.903,33.715c2.043,0 3.555,0.74 4.536,2.222c0.776,1.234 1.185,3.004 1.185,5.308c0,2.305 -0.449,4.074 -1.308,5.308c-0.98,1.317 -2.533,2.017 -4.576,2.017c-1.512,0 -2.738,-0.617 -3.719,-1.852c-1.103,-1.358 -1.634,-3.168 -1.634,-5.391l0,-0.329c0,-2.14 0.409,-3.827 1.308,-5.102c0.98,-1.482 2.37,-2.181 4.208,-2.181Z" style={{ fill: '#535353', fillRule: 'nonzero' }} /><rect x="319.02" y="21.822" width="4.658" height="29.957" style={{ fill: '#535353', fillRule: 'nonzero' }} /><path d="M336.918,29.929c-3.229,0 -5.803,1.029 -7.723,3.168c-1.962,2.099 -2.902,4.774 -2.902,8.066c0,3.251 0.94,5.925 2.861,7.983c1.961,2.14 4.535,3.21 7.764,3.21c3.187,0 5.802,-1.07 7.763,-3.21c1.88,-2.058 2.861,-4.732 2.861,-7.983c0,-3.292 -0.981,-5.967 -2.902,-8.066c-1.961,-2.139 -4.535,-3.168 -7.722,-3.168ZM336.918,33.756c1.92,0 3.432,0.741 4.494,2.304c0.899,1.317 1.39,3.045 1.39,5.103c0,2.016 -0.491,3.703 -1.39,5.061c-1.062,1.523 -2.574,2.305 -4.494,2.305c-1.962,0 -3.433,-0.782 -4.495,-2.305c-0.899,-1.317 -1.349,-3.004 -1.349,-5.061c0,-2.058 0.45,-3.786 1.349,-5.103c1.062,-1.563 2.533,-2.304 4.495,-2.304Z" style={{ fill: '#535353', fillRule: 'nonzero' }} /><path d="M359.719,29.929c-1.308,0 -2.411,0.37 -3.351,1.152c-0.777,0.576 -1.43,1.399 -2.002,2.469l0,-3.045l-4.618,0l0,21.274l4.618,0l0,-11.275c0,-1.81 0.531,-3.25 1.675,-4.361c0.981,-0.988 2.084,-1.441 3.31,-1.441c0.98,0 1.961,0.124 3.024,0.453l0,-4.691c-0.736,-0.37 -1.635,-0.535 -2.656,-0.535Z" style={{ fill: '#535353', fillRule: 'nonzero' }} /><path d="M372.141,29.929c-3.147,0 -5.639,1.07 -7.437,3.251c-1.88,2.14 -2.779,4.773 -2.779,7.942c0,3.539 0.981,6.296 2.942,8.312c1.839,1.934 4.331,2.922 7.478,2.922c2.819,0 5.148,-0.823 7.028,-2.428c1.471,-1.317 2.452,-3.004 2.942,-4.979l-4.658,0c-0.572,1.193 -1.185,2.057 -1.88,2.592c-0.899,0.658 -2.043,0.988 -3.473,0.988c-1.675,0 -2.983,-0.535 -3.882,-1.564c-0.899,-1.029 -1.43,-2.551 -1.594,-4.527l15.814,0c0,-3.826 -0.858,-6.83 -2.533,-8.97c-1.839,-2.387 -4.495,-3.539 -7.968,-3.539ZM372.263,33.756c3.269,0 5.108,1.687 5.517,5.144l-10.87,0c0.286,-1.688 0.858,-2.963 1.716,-3.827c0.899,-0.906 2.084,-1.317 3.637,-1.317Z" style={{ fill: '#535353', fillRule: 'nonzero' }} /><path d="M394.41,29.929c-1.307,0 -2.411,0.37 -3.35,1.152c-0.777,0.576 -1.431,1.399 -2.003,2.469l0,-3.045l-4.617,0l0,21.274l4.617,0l0,-11.275c0,-1.81 0.532,-3.25 1.676,-4.361c0.98,-0.988 2.084,-1.441 3.31,-1.441c0.98,0 1.961,0.124 3.023,0.453l0,-4.691c-0.735,-0.37 -1.634,-0.535 -2.656,-0.535Z" style={{ fill: '#535353', fillRule: 'nonzero' }} /></g></g></svg>
+);
 
 export default Logo;
